ADD Trigger Handling

+++ src/main/java/com/replace/replace/api/dynamic/kernel/entry/Delete.java
@@ -1,10 +1,13 @@
 package com.replace.replace.api.dynamic.kernel.entry;
 
 import com.replace.replace.api.dynamic.kernel.entity.EntityHandler;
+import com.replace.replace.api.dynamic.kernel.router.RouteHandler;
 import com.replace.replace.api.history.HistoryHandler;
 import com.replace.replace.api.request.Request;
 import com.replace.replace.repository.DefaultRepository;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
 
 /**
  * @author Romain Lavabre <romainlavabre98@gmail.com>
@@ -21,13 +24,35 @@ public class Delete implements DeleteEntry {
 
 
     @Override
-    public void delete( Request request, Object subject ) {
+    public void delete( Request request, Object subject, RouteHandler.Route route )
+            throws Throwable {
+        delete( request, subject, route.getTriggers() );
+    }
+
+
+    protected void delete( Request request, Object subject, List< com.replace.replace.api.dynamic.annotation.Trigger > nextTriggers )
+            throws Throwable {
+        delete( request, subject, nextTriggers, null );
+    }
+
+
+    protected void delete( Request request, Object subject, List< com.replace.replace.api.dynamic.annotation.Trigger > nextTriggers, Object executor )
+            throws Throwable {
+
+        if ( executor != null ) {
+            (( com.replace.replace.api.crud.Delete ) executor).delete( request, subject );
+            return;
+        }
+
         EntityHandler.Entity entity = EntityHandler.getEntity( subject.getClass() );
+
+        historyHandler.delete( subject );
+
 
         DefaultRepository defaultRepository = entity.getDefaultRepository();
 
         defaultRepository.remove( subject );
 
-        historyHandler.delete( subject );
+        Trigger.getInstance().handleTriggers( request, nextTriggers, subject );
     }
 }
